@@ -22,15 +22,16 @@
         }
     ],
     "require": {
-        "php": "^8.0",
-        "league/flysystem": "^1.0",
+        "php": "^8.1",
+        "league/flysystem": "^3.0",
         "microsoft/microsoft-graph": "^1.0"
     },
     "require-dev": {
         "friendsofphp/php-cs-fixer": "^3.10.0",
         "phpstan/phpstan-deprecation-rules": "^1.0.0",
         "phpstan/phpstan": "^1.8.2",
-        "phpunit/phpunit": "^9.5.23"
+        "phpunit/phpunit": "^9.5.23",
+        "symfony/var-dumper": "^6.1"
     },
     "scripts": {
         "ci": [
